requirements.txt
loguru~=0.7.0
Pillow~=9.5.0